some compilers apparently don't do the default int->double conversion so we must cast

+++ src/StatHist.c
@@ -243,7 +243,7 @@ Null(double x)
 void
 statHistEnumInit(StatHist * H, int last_enum)
 {
-    statHistInit(H, last_enum + 3, &Null, &Null, -1, last_enum + 1 + 1);
+    statHistInit(H, last_enum + 3, &Null, &Null, (double) -1, (double) (last_enum + 1 + 1));
 }
 
 void
